[type:refactor] Put the creation logic of HttpClient in a separate class
