+++ shenyu-spring-boot-starter/shenyu-spring-boot-starter-plugin/shenyu-spring-boot-starter-plugin-httpclient/src/main/java/org/apache/shenyu/springboot/starter/plugin/httpclient/HttpClientPluginConfiguration.java
@@ -17,38 +17,23 @@
 
 package org.apache.shenyu.springboot.starter.plugin.httpclient;
 
-import io.netty.channel.ChannelOption;
-import io.netty.handler.ssl.util.InsecureTrustManagerFactory;
-import io.netty.handler.timeout.IdleStateHandler;
-import io.netty.handler.timeout.ReadTimeoutHandler;
-import io.netty.handler.timeout.WriteTimeoutHandler;
-import org.apache.commons.lang3.ArrayUtils;
-import org.apache.commons.lang3.StringUtils;
 import org.apache.shenyu.plugin.api.ShenyuPlugin;
 import org.apache.shenyu.plugin.httpclient.NettyHttpClientPlugin;
 import org.apache.shenyu.plugin.httpclient.WebClientPlugin;
 import org.apache.shenyu.plugin.httpclient.config.HttpClientProperties;
-import org.apache.shenyu.plugin.httpclient.config.HttpClientProperties.Pool;
 import org.springframework.beans.factory.ObjectProvider;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.boot.context.properties.PropertyMapper;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.client.reactive.ReactorClientHttpConnector;
 import org.springframework.web.reactive.function.client.ExchangeStrategies;
 import org.springframework.web.reactive.function.client.WebClient;
 import reactor.netty.http.client.HttpClient;
-import reactor.netty.resources.ConnectionProvider;
 import reactor.netty.resources.LoopResources;
-import reactor.netty.tcp.DefaultSslContextSpec;
-import reactor.netty.tcp.SslProvider;
-import reactor.netty.transport.ProxyProvider;
 
-import java.security.cert.X509Certificate;
-import java.time.Duration;
 import java.util.Objects;
-import java.util.concurrent.TimeUnit;
 
 /**
  * The type Http client plugin configuration.
@@ -89,132 +74,12 @@ public class HttpClientPluginConfiguration {
      * @return the http client
      */
     @Bean
-    public HttpClient httpClient(final HttpClientProperties properties,
+    @ConditionalOnMissingBean({HttpClient.class, HttpClientFactory.class})
+    public HttpClientFactory httpClient(final HttpClientProperties properties,
                                  final ObjectProvider<LoopResources> provider) {
-        // configure pool resources.
-        HttpClientProperties.Pool pool = properties.getPool();
-        ConnectionProvider connectionProvider = buildConnectionProvider(pool);
-        HttpClient httpClient = HttpClient.create(connectionProvider)
-                .option(ChannelOption.CONNECT_TIMEOUT_MILLIS, properties.getConnectTimeout());
-        HttpClientProperties.Proxy proxy = properties.getProxy();
-        if (StringUtils.isNotEmpty(proxy.getHost())) {
-            httpClient = setHttpClientProxy(httpClient, proxy);
-        }
-        httpClient.doOnConnected(connection -> {
-            connection.addHandlerLast(new IdleStateHandler(properties.getReaderIdleTime(), properties.getWriterIdleTime(), properties.getAllIdleTime(), TimeUnit.MILLISECONDS));
-            connection.addHandlerLast(new WriteTimeoutHandler(properties.getWriteTimeout(), TimeUnit.MILLISECONDS));
-            connection.addHandlerLast(new ReadTimeoutHandler(properties.getReadTimeout(), TimeUnit.MILLISECONDS));
-        });
-        final LoopResources loopResources = provider.getIfAvailable();
-        if (Objects.nonNull(loopResources)) {
-            httpClient.runOn(loopResources);
-        }
-        HttpClientProperties.Ssl ssl = properties.getSsl();
-        if (StringUtils.isNotEmpty(ssl.getKeyStorePath())
-                || ArrayUtils.isNotEmpty(ssl.getTrustedX509CertificatesForTrustManager())
-                || ssl.isUseInsecureTrustManager()) {
-            httpClient = httpClient.secure(sslContextSpec -> setSsl(sslContextSpec, ssl));
-        }
-        if (properties.isWiretap()) {
-            httpClient = httpClient.wiretap(true);
-        }
-        // set to false, fix java.io.IOException: Connection reset by peer
-        // see https://github.com/reactor/reactor-netty/issues/388
-        return httpClient.keepAlive(properties.isKeepAlive());
+        return new HttpClientFactory(properties, provider.getIfAvailable());
     }
 
-    /**
-     * set http proxy.
-     * @param httpClient http client
-     * @param proxy proxy
-     * @return HttpClient
-     */
-    private HttpClient setHttpClientProxy(final HttpClient httpClient, final HttpClientProperties.Proxy proxy) {
-        return httpClient.proxy(proxySpec -> {
-            ProxyProvider.Builder builder = proxySpec
-                    .type(ProxyProvider.Proxy.HTTP)
-                    .host(proxy.getHost());
-            PropertyMapper map = PropertyMapper.get();
-            map.from(proxy::getPort).whenNonNull().to(builder::port);
-            map.from(proxy::getUsername).whenHasText()
-                    .to(builder::username);
-            map.from(proxy::getPassword).whenHasText()
-                    .to(password -> builder.password(s -> password));
-            map.from(proxy::getNonProxyHostsPattern).whenHasText()
-                    .to(builder::nonProxyHosts);
-        });
-    }
-
-    private void setSsl(final SslProvider.SslContextSpec sslContextSpec, final HttpClientProperties.Ssl ssl) {
-        SslProvider.ProtocolSslContextSpec spec = DefaultSslContextSpec.forClient()
-                .configure(sslContextBuilder -> {
-                    X509Certificate[] trustedX509Certificates = ssl.getTrustedX509CertificatesForTrustManager();
-                    if (ArrayUtils.isNotEmpty(trustedX509Certificates)) {
-                        sslContextBuilder.trustManager(trustedX509Certificates);
-                    } else if (ssl.isUseInsecureTrustManager()) {
-                        sslContextBuilder.trustManager(InsecureTrustManagerFactory.INSTANCE);
-                    }
-                    sslContextBuilder.keyManager(ssl.getKeyManagerFactory());
-                    sslContextBuilder.sslProvider(ssl.getDefaultConfigurationType());
-                });
-        sslContextSpec.sslContext(spec)
-                .handshakeTimeout(ssl.getHandshakeTimeout())
-                .closeNotifyFlushTimeout(ssl.getCloseNotifyFlushTimeout())
-                .closeNotifyReadTimeout(ssl.getCloseNotifyReadTimeout());
-    }
-
-    private ConnectionProvider buildConnectionProvider(final HttpClientProperties.Pool pool) {
-        ConnectionProvider connectionProvider;
-        if (pool.getType() == HttpClientProperties.Pool.PoolType.DISABLED) {
-            connectionProvider = ConnectionProvider.newConnection();
-        } else if (pool.getType() == HttpClientProperties.Pool.PoolType.FIXED) {
-            // reactor remove fixed pool by fixed method from 0.9.4
-            // reason: https://github.com/reactor/reactor-netty/issues/1499 and https://github.com/reactor/reactor-netty/issues/1960
-            connectionProvider = buildFixedConnectionPool(pool);
-        } else {
-            // please see https://projectreactor.io/docs/netty/release/reference/index.html#_connection_pool_2
-            // reactor remove elastic pool by elastic method from 0.9.4
-            // reason: https://github.com/reactor/reactor-netty/issues/1499 and https://github.com/reactor/reactor-netty/issues/1960
-            connectionProvider = buildElasticConnectionPool(pool);
-        }
-        return connectionProvider;
-    }
-
-    /**
-     * build fixed connection pool.
-     *
-     * @param pool connection pool params
-     * @return {@link ConnectionProvider}
-     */
-    public static ConnectionProvider buildFixedConnectionPool(final Pool pool) {
-        if (pool.getMaxConnections() <= 0) {
-            throw new IllegalArgumentException("Max Connections value must be strictly positive");
-        }
-        if (pool.getAcquireTimeout() < 0) {
-            throw new IllegalArgumentException("Acquire Timeout value must be positive");
-        }
-        ConnectionProvider.Builder builder = ConnectionProvider.builder(pool.getName())
-                .maxConnections(pool.getMaxConnections())
-                .pendingAcquireTimeout(Duration.ofMillis(pool.getAcquireTimeout()))
-                .maxIdleTime(pool.getMaxIdleTime());
-        return builder.build();
-    }
-
-    /**
-     * build elastic connection provider pool.
-     *
-     * @param pool connection pool params
-     * @return {@link ConnectionProvider} elastic pool
-     */
-    public ConnectionProvider buildElasticConnectionPool(final Pool pool) {
-        // about the args, please see https://projectreactor.io/docs/netty/release/reference/index.html#_connection_pool_2
-        ConnectionProvider.Builder builder = ConnectionProvider.builder(pool.getName())
-                .maxConnections(Integer.MAX_VALUE)
-                .pendingAcquireTimeout(Duration.ofMillis(0))
-                .pendingAcquireMaxCount(-1)
-                .maxIdleTime(pool.getMaxIdleTime());
-        return builder.build();
-    }
 
     /**
      * The type Web client configuration.
